test: :white_check_mark: E2E tests and minor changes in services and views
- Added all e2e tests and their configuration - Fixed config for correct creation of mongo uris - Added some tests ids and labels for tests

+++ webapp/src/components/PrivateRoute.js
@@ -4,10 +4,14 @@ import PropTypes from "prop-types";
 import { SessionContext } from "../SessionContext";
 
 // Private route component to restrict access
-const PrivateRoute = ({ element, roles = [] }) => {
+const PrivateRoute = ({ element, roles = [], logged = false }) => {
   const { isLoggedIn, role, destroySession } = useContext(SessionContext); // Check if user has an active session;
 
+  if (logged && isLoggedIn === true) {
+    return <Navigate to="/" />; // Redirect to home if already logged in
+  }
   if (!isLoggedIn) {
+    if (logged === true) return element; // Allow access to login/forgot password/reset password if not logged in
     destroySession();
     return <Navigate to="/login" />; // Redirect to login if no session
   }
